pyproject.toml
[project]
name = "printer-queue-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose (==3.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (==2.8.0)",
    "aioredis (==2.0.1)",
    "redis (>=4.2,<5.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "minio (>=7.2.18,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
